add options to pureftpd-ldap.conf to allow overriding LDAP uid/gid values
This patch adds to two options to pureftpd-ldap.conf:

    LDAPForceDefaultUID True
    LDAPForceDefaultGID True

These options both default to `False`. Any value other than `True` (case
insensitive) is also treated as `False`.  When set these options cause the
respective uid or gid value returned by the LDAP server for a username to be
ignored and instead use the value set by `LDAPDefaultUID` or `LDAPDefaultGID`.
If the appropriate `LDAPDefaultXID` option is not set, these options have no
effect.

This is useful for allowing users to authenticate against LDAP but
access or create content with a common set of ownership/permissions.  It
also provides a measure of security in that it prevents pure-ftpd
processes from being created with arbitrary uid/gids that may conflict
with local accounts.

An example complete configuration snippet:

    LDAPDefaultUID 777
    LDAPForceDefaultUID True
    LDAPDefaultGID 888
    LDAPForceDefaultGID True