dopass: Allow chaining of chroot for homedir.
Chaining of chroot is useful when parts of the homedir are supplied by untrusted users
or untrusted users can possibly create symlinks in the path to homedir.

Each point in the chroot chain allows locking the user into a certain directory while
further steps allow additional restrictions without possibility to widen the access
by .. components or symlinks in the homedir value.