Allow client to gracefully fail-over if TLS session has not been established yet
Instead of disconnecting the client when a TLS session has not been established yet, allow the client to gracefully "fail-over" to an encrypted session (by sending an AUTH TLS command). This behavior also complies with the error code (421) which belongs to "4xx Transient Negative Completion reply" category, which indicates that the error condition is temporary.